cmd/geth: resolve errors caused by merge with 119fe1dd
Moves core-geth-specific tests to non-conflicting
file.

Date: 2021-03-21 07:46:25-05:00
